选择支付方式弹窗

+++ cxpay/src/main/java/com/hm/cxpay/bean/UserBean.java
@@ -7,8 +7,8 @@ public class UserBean {
     private String realName = "";//真实姓名
     private String cardId = "";//身份证号码
 
-    private int balance = 0;//余额
-    private int id = 0;//用户id
+    private long balance = 0;//余额
+    private long id = 0;//用户id
     private String identityNo = "";//证件号码
     private String identityType = "";//证件类型 1身份证 2护照
     private int payPwdStat = 0;//支付密码设置状态 1:已设置, 其他情况未设置
@@ -47,19 +47,19 @@ public class UserBean {
         this.cardId = cardId;
     }
 
-    public int getBalance() {
+    public long getBalance() {
         return balance;
     }
 
-    public void setBalance(int balance) {
+    public void setBalance(long balance) {
         this.balance = balance;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
